media-libs/kvazaar: fix AVX2 flags on GCC
On older CPUS without support for popcnt or lzcnt compilation of
kvazaar fails due to these instructions being used by the AVX2
code but the relevant flags to support them not being set by
the Makefile.

Upstream fixed this in commit
99423a825027c862bbfd50e112a1846a013b01a7

Backport this patch so compilation works again on systems
without CPU support for these instructions.

Closes: https://bugs.gentoo.org/739776
Bug: https://github.com/ultravideo/kvazaar/pull/301
Signed-off-by: Francisco Blas (klondike) Izquierdo Riera <klondike@gentoo.org>